+++ tests/peer/media/mod.rs
@@ -203,6 +203,7 @@ mod sender_patch {
             receivers: None,
             media_direction: Some(MediaDirection::RecvOnly),
             muted: None,
+            encoding_parameters: None,
         });
         sender.state().when_updated().await;
 
@@ -217,6 +218,7 @@ mod sender_patch {
             receivers: None,
             media_direction: Some(MediaDirection::RecvOnly),
             muted: None,
+            encoding_parameters: None,
         });
         sender.state().when_updated().await;
 
@@ -231,6 +233,7 @@ mod sender_patch {
             receivers: None,
             media_direction: Some(MediaDirection::SendRecv),
             muted: None,
+            encoding_parameters: None,
         });
         sender.state().when_updated().await;
 
@@ -245,6 +248,7 @@ mod sender_patch {
             receivers: None,
             media_direction: Some(MediaDirection::RecvOnly),
             muted: None,
+            encoding_parameters: None,
         });
         sender.state().when_updated().await;
         assert!(sender.general_disabled());
@@ -254,6 +258,7 @@ mod sender_patch {
             receivers: None,
             media_direction: Some(MediaDirection::RecvOnly),
             muted: None,
+            encoding_parameters: None,
         });
         sender.state().when_updated().await;
 
@@ -268,6 +273,7 @@ mod sender_patch {
             receivers: None,
             media_direction: None,
             muted: None,
+            encoding_parameters: None,
         });
         sender.state().when_updated().await;
 
@@ -344,6 +350,7 @@ mod receiver_patch {
             receivers: None,
             media_direction: Some(MediaDirection::RecvOnly),
             muted: None,
+            encoding_parameters: None,
         });
         receiver.state().when_updated().await;
 
@@ -358,6 +365,7 @@ mod receiver_patch {
             receivers: None,
             media_direction: Some(MediaDirection::RecvOnly),
             muted: None,
+            encoding_parameters: None,
         });
         receiver.state().when_updated().await;
 
@@ -372,6 +380,7 @@ mod receiver_patch {
             receivers: None,
             media_direction: Some(MediaDirection::SendRecv),
             muted: None,
+            encoding_parameters: None,
         });
         receiver.state().when_updated().await;
 
@@ -386,6 +395,7 @@ mod receiver_patch {
             receivers: None,
             media_direction: Some(MediaDirection::RecvOnly),
             muted: None,
+            encoding_parameters: None,
         });
         receiver.state().when_updated().await;
         assert!(!receiver.enabled_general());
@@ -395,6 +405,7 @@ mod receiver_patch {
             receivers: None,
             media_direction: Some(MediaDirection::RecvOnly),
             muted: None,
+            encoding_parameters: None,
         });
         receiver.state().when_updated().await;
 
@@ -409,6 +420,7 @@ mod receiver_patch {
             receivers: None,
             media_direction: None,
             muted: None,
+            encoding_parameters: None,
         });
         receiver.state().when_updated().await;